refactor(carousel): move ui utils to lib
