+++ mobile/kmp/ui/src/commonMain/kotlin/com/atls/hyperion/ui/components/carousel/lib/CenterOfItem.kt
@@ -1,4 +1,4 @@
-package com.atls.hyperion.ui.components.carousel.ui
+package com.atls.hyperion.ui.components.carousel.lib
 
 fun centerOfItem(offset: Int, size: Int): Int =
     offset + size / 2
