+++ mobile/kmp/ui/src/commonMain/kotlin/com/atls/hyperion/ui/components/carousel/lib/CenterOfViewport.kt
@@ -1,4 +1,4 @@
-package com.atls.hyperion.ui.components.carousel.ui
+package com.atls.hyperion.ui.components.carousel.lib
 
 fun centerOfViewport(viewportStartOffset: Int, viewportWidth: Int): Int =
     viewportStartOffset + viewportWidth / 2
