+++ mobile/kmp/ui/src/commonMain/kotlin/com/atls/hyperion/ui/components/carousel/lib/NormalizedStartIndex.kt
@@ -1,4 +1,4 @@
-package com.atls.hyperion.ui.components.carousel.ui
+package com.atls.hyperion.ui.components.carousel.lib
 
 fun normalizedStartIndex(centerIndex: Int, itemsSize: Int): Int {
     val middle = Int.MAX_VALUE / 2
